code cleanup, split up files
