--- src/powerquery-parser/inspection/scope/scopeItem.ts
@@ -1,63 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-import { Ast, Constant } from "../../language";
-import { TXorNode } from "../../parser";
-
-export type TScopeItem =
-    | EachScopeItem
-    | LetVariableScopeItem
-    | ParameterScopeItem
-    | RecordFieldScopeItem
-    | SectionMemberScopeItem
-    | UndefinedScopeItem;
-
-export type TKeyValuePairScopeItem = LetVariableScopeItem | RecordFieldScopeItem | SectionMemberScopeItem;
-
-export const enum ScopeItemKind {
-    Each = "Each",
-    LetVariable = "LetVariable",
-    Parameter = "Parameter",
-    RecordField = "RecordField",
-    SectionMember = "SectionMember",
-    Undefined = "Undefined",
-}
-
-export interface IScopeItem {
-    readonly kind: ScopeItemKind;
-    readonly id: number;
-    readonly isRecursive: boolean;
-}
-
-export interface IKeyValuePairScopeItem<
-    Key extends Ast.Identifier | Ast.GeneralizedIdentifier,
-    Kind extends ScopeItemKind.LetVariable | ScopeItemKind.RecordField | ScopeItemKind.SectionMember
-> extends IScopeItem {
-    readonly kind: Kind;
-    readonly key: Key;
-    readonly maybeValue: TXorNode | undefined;
-}
-
-export interface EachScopeItem extends IScopeItem {
-    readonly kind: ScopeItemKind.Each;
-    readonly eachExpression: TXorNode;
-}
-
-export type LetVariableScopeItem = IKeyValuePairScopeItem<Ast.Identifier, ScopeItemKind.LetVariable>;
-
-export interface ParameterScopeItem extends IScopeItem {
-    readonly kind: ScopeItemKind.Parameter;
-    readonly name: Ast.Identifier;
-    readonly isOptional: boolean;
-    readonly isNullable: boolean;
-    readonly maybeType: Constant.PrimitiveTypeConstantKind | undefined;
-}
-
-export type RecordFieldScopeItem = IKeyValuePairScopeItem<Ast.GeneralizedIdentifier, ScopeItemKind.RecordField>;
-
-export type SectionMemberScopeItem = IKeyValuePairScopeItem<Ast.Identifier, ScopeItemKind.SectionMember>;
-
-export interface UndefinedScopeItem extends IScopeItem {
-    readonly kind: ScopeItemKind.Undefined;
-    readonly xorNode: TXorNode;
-}
